Add files via upload
HTML, CSS and JavaScript have been used in this. Mainly it has 3 basics currency which are often used in the market as compared to other currency such as Indian Rupees, US Doller, Euro.
You can select it. How many currency converter or which one from the basics currency was taken and then check the running conversion rate of that currency you select , in many other countries .